"remember me" feature

+++ ims/src/main/java/com/khoilnm/ims/serviceImpl/AuthenticationServiceImpl.java
@@ -1,10 +1,8 @@
 package com.khoilnm.ims.serviceImpl;
 
-import com.khoilnm.ims.common.ConstantUtils;
-import com.khoilnm.ims.model.User;
+import com.khoilnm.ims.exceptions.TokenRevocationException;
 import com.khoilnm.ims.service.AuthenticationService;
 import com.khoilnm.ims.service.JwtService;
-import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.extern.slf4j.Slf4j;
@@ -12,6 +10,7 @@ import org.springframework.context.annotation.Lazy;
 import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.Authentication;
+import org.springframework.security.core.AuthenticationException;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
@@ -20,15 +19,17 @@ import org.springframework.stereotype.Service;
 public class AuthenticationServiceImpl implements AuthenticationService {
     private final AuthenticationManager authenticationManager;
     private final JwtService jwtService;
+    private final TokenRevocationServiceImpl tokenRevocationService;
 
-    public AuthenticationServiceImpl(@Lazy AuthenticationManager authenticationManager, JwtService jwtService) {
+    public AuthenticationServiceImpl(@Lazy AuthenticationManager authenticationManager, JwtService jwtService, TokenRevocationServiceImpl tokenRevocationService) {
         this.authenticationManager = authenticationManager;
         this.jwtService = jwtService;
+        this.tokenRevocationService = tokenRevocationService;
     }
 
     /**
-     * @param email
-     * @param password
+     * @param email    String
+     * @param password String
      * @return Authentication
      */
     @Override
@@ -37,44 +38,72 @@ public class AuthenticationServiceImpl implements AuthenticationService {
     }
 
     /**
-     * @param authentication
-     * @return User
-     */
-    @Override
-    public User get_user_info(Authentication authentication) {
-        return (User) authentication.getPrincipal();
-    }
-
-    /**
-     * @param request
-     * @param response
-     * @param authentication
+     * @param request        HttpServletRequest
+     * @param response       HttpServletResponse
+     * @param authentication Authentication
      */
     @Override
     public void logout(HttpServletRequest request, HttpServletResponse response, Authentication authentication) {
         try {
+            if (authentication == null) return;
             final String refreshToken = jwtService.getRefreshTokenFromCookie(request);
-            String email = jwtService.getEmailFromRefreshToken(refreshToken);
+            String email = jwtService.getEmailFromToken(refreshToken);
             jwtService.removeByEmail(email);
+            tokenRevocationService.revokeToken(refreshToken, TokenRevocationServiceImpl.RevokeReason.USER_LOGOUT);
         } catch (Exception e) {
             log.error("Error while logging out: {}", e.getMessage());
-            throw new RuntimeException(e);
+            throw new RuntimeException(e.getMessage());
+        } catch (TokenRevocationException e) {
+            log.error("Error while revoking token to logout: {}", e.getMessage());
+            throw new RuntimeException(e.getMessage());
         } finally {
-            Cookie accessCookie = new Cookie(ConstantUtils.JWT_ACCESS_TOKEN, null);
-            accessCookie.setPath("/");
-            accessCookie.setHttpOnly(true);
-            accessCookie.setMaxAge(0);
-            response.addCookie(accessCookie);
+
+            jwtService.removeAccessTokenCookie(response);
 
             // Clear refresh token cookie
-            Cookie refreshTokenCookie = new Cookie(ConstantUtils.JWT_REFRESH_TOKEN, null);
-            refreshTokenCookie.setPath("/");
-            refreshTokenCookie.setHttpOnly(true);
-            refreshTokenCookie.setMaxAge(0);
-            response.addCookie(refreshTokenCookie);
+            jwtService.removeRefreshTokenCookie(response);
 
             // Clear security context
             SecurityContextHolder.clearContext();
         }
+    }
+
+    /**
+     * <h1>Revoke all tokens and force logout user when password is changed</h1>
+     *
+     * @param email       String
+     * @param newPassword String
+     * @throws TokenRevocationException Exception
+     */
+    public void changePassword(String email, String newPassword) throws TokenRevocationException {
+        // TODO: Implement password change
+
+
+        // Revoke all tokens
+        tokenRevocationService.revokeAllUserTokens(email, TokenRevocationServiceImpl.RevokeReason.PASSWORD_CHANGE);
+    }
+
+    /**
+     * <h1>Revoke all tokens and force logout user when account is disabled</h1>
+     *
+     * @param email String
+     * @throws TokenRevocationException Exception
+     */
+    public void disableAccount(String email) throws TokenRevocationException {
+        // TODO: Implement account disable
+
+        // Revoke all tokens
+        tokenRevocationService.revokeAllUserTokens(email, TokenRevocationServiceImpl.RevokeReason.ACCOUNT_DISABLED);
+    }
+
+    /**
+     * <h1>Revoke all tokens when user forcing to log out on all devices</h1>
+     *
+     * @param email String
+     * @throws TokenRevocationException Exception
+     */
+    public void forceLogout(String email) throws TokenRevocationException {
+        // Revoke all tokens
+        tokenRevocationService.revokeAllUserTokens(email, TokenRevocationServiceImpl.RevokeReason.FORCE_LOGOUT);
     }
 }
